% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/current_download.R
\name{current_download}
\alias{current_download}
\title{Download the latest data from Meteoclimatic}
\usage{
current_download(id = "ESCAT080000000", save_excel = TRUE)
}
\arguments{
\item{id}{character. Id of the CCAA, province or station, used by Meteoclimatic. Default: "ESCAT080000000". Province of Barcelona.}

\item{save_excel}{logical. Do you want to save the downloaded data in an Excel file?}
}
\value{
A list with: \itemize{
   \item{A data.frame containing georreferenced daily meteorological values for all the stations located in the requested id
}}
}
\description{
\code{current_download} allows to download all the latest daily meteorological from the Meteoclimatic network.
}
\examples{
latest_data <- current_download(id = "ESCAT080000000", save_excel = FALSE)

}
\seealso{
\code{\link{historical_download}}
}
