% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historical_download.R
\name{historical_download}
\alias{historical_download}
\title{Download daily historical records from Meteoclimatic}
\usage{
historical_download(
  id = "ESIBA",
  dates = c("2021-05-16", "2021-07-15"),
  months = c(1:12),
  save_excel = TRUE
)
}
\arguments{
\item{id}{character. Id of the CCAA, used by Meteoclimatic. Default "ESIBA" Balearic Islands.}

\item{dates}{character. Starting and ending date. Alternatively, a unique date is also allowed.}

\item{months}{numeric. Optional. Month subset Only applies for ranges of dates.}

\item{save_excel}{logical. Save the downloaded database in an Excel file}
}
\value{
A list with: \itemize{
   \item{A data.frame containing georreferenced daily meteorological values for all the stations located in the requested CCAA.
}}
}
\description{
\code{historical_download} allows to download all the historical
daily Meteoclimatic database. Records starts in 2012.
}
\examples{
historical_data <- historical_download(id = "ESIBA",
dates = c("2021-05-16","2021-07-15"),
save_excel = FALSE)

}
\seealso{
\code{\link{current_download}}
}
